Changing hostname for database connection
I was looking through my dao from 421 and I think the problem we are having is with the host name, we are running the host as 127.0.0.1 which is for local hosting, but I believe for the website to work it may need to be the actual hostname i.e. where we are connecting to so it may nee dto be 132.178.215.87 for now and when we are done we will need to switch it to taticketing.boisestate.edu so it cannot be accessed outside of the intranet.